refactor: use the `isAuth` flag to fetch courses
