+++ src/app/api/courses/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from 'next/server';
-import { getAllCourses } from '~/services/courses/courses';
+import { getAllCourses } from '~/services/courses/get-courses';
 
 export async function GET() {
   const courses = await getAllCourses();
